changes in order and models

+++ ondoc/api/v1/insurance/serializers.py
@@ -89,7 +89,7 @@ class InsuranceTransactionModelSerializer(serializers.Serializer):
     insurance_plan = serializers.PrimaryKeyRelatedField(queryset=InsurancePlans.objects.all())
     user = serializers.PrimaryKeyRelatedField(queryset=User.objects.all())
     # order = serializers.PrimaryKeyRelatedField(queryset=account_models.Order.objects.all())
-    amount = serializers.DecimalField(max_digits=10, decimal_places=2)
+    amount = serializers.IntegerField()
     status_type = serializers.CharField(max_length=50)
     # insured_members = JSONField(null=True, blank=True)
     # insured_members = serializers.ListField(child=serializers.PrimaryKeyRelatedField(queryset=InsuredMembers.objects.all))
